Version 3.2
Enabled new styling for the time page.  Links NOT enabled yet.

NOTE:
Running an external CSS file wasn't working (the CSS wasn't being
applied) so the CSS was embedded internally in the time.html file and is
running fine now.  This isn't ideal, but requirements don't state that
the CSS needs to be external and numerous workaround attempts at
externalizing the CSS have failed, so it's being ignored for now.