Version 3.3.
Links on time page are now supported.

Added a new page "localhost:(xxx)/menu" which simply displays the menu
consisting of "Home", "Time", "Logout", and "About Us".

Templates/HTML files have been moved into a sub-folder called
"Templates" and their embedded paths in timeserver.go now require them
to be in such a directory which is a child of the directory
timeserver.go is located in.