Version 3.3.2
Added templates flag.  Add path to templates folder relative to
timeserver.go file with templates flag to change the directory HTML
files are being read from.  The default path is "Templates/"

Also added a redirect for the logout page if the user is not logged in.
If user is logged in, the logout page functions identically to before.
If the user is NOT logged in, the logout page will redirect the user to
the login page in the same manner the "/" and "/index.html" pages
redirect the user.